Refactored: - Coop Record Flow - Chicken Transfer Flow - Brooder Record Flow
