+++ db/database.js
@@ -1,0 +1,10 @@
+import mysql2 from 'mysql2';
+
+const pool = mysql2.createPool({
+  host: process.env.MYSQL_HOST,
+  user: process.env.MYSQL_USER,
+  password: process.env.MYSQL_PASSWORD,
+  database: process.env.MYSQL_DATABASE
+}).promise();
+
+export default pool;
