+++ views/create-brooder-record.ejs
@@ -145,7 +145,7 @@
 
 
     <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-      <form class="needs-validation" action="/submit-brooder-record" method="post" role="form" class="text-start">
+      <form class="needs-validation" action="/brooder/submit" method="post" role="form" class="text-start">
         <div id="formRepeat">
           <div class="row border px-2 py-2 mx-1">
             <div class="row mx-1">
