+++ views/create-coop-record.ejs
@@ -144,7 +144,7 @@
 
 
     <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-      <form class="needs-validation" action="/submit-coop-record" method="post" role="form" class="text-start">
+      <form class="needs-validation" action="/coop/submit" method="post" role="form" class="text-start">
         <div id="formRepeat">
           <div class="row border px-2 py-2 mx-1">
             <div class="row mx-1">
